# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chop Plan API - prepaid lunch subscription service for Lagos restaurants
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UserSignupInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  /** Lagos area/neighborhood */
  area: string;
}

export interface VendorSignupInput {
  businessName: string;
  ownerName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  area: string;
  cuisineType: string;
  description?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type AuthResponseRole = typeof AuthResponseRole[keyof typeof AuthResponseRole];


export const AuthResponseRole = {
  user: 'user',
  vendor: 'vendor',
  admin: 'admin',
} as const;

export interface AuthResponse {
  token: string;
  role: AuthResponseRole;
  id: number;
  name: string;
  email: string;
}

export type MeResponseRole = typeof MeResponseRole[keyof typeof MeResponseRole];


export const MeResponseRole = {
  user: 'user',
  vendor: 'vendor',
  admin: 'admin',
} as const;

export interface MeResponse {
  id: number;
  name: string;
  email: string;
  role: MeResponseRole;
  /** @nullable */
  area?: string | null;
}

export interface Vendor {
  id: number;
  businessName: string;
  area: string;
  cuisineType: string;
  coverImage: string;
  rating: number;
  subscriberCount: number;
  /** Lowest monthly plan price in Naira */
  lowestPlanPrice: number;
  /** @nullable */
  description?: string | null;
}

export interface SubscriptionPlan {
  id: number;
  name: string;
  daysPerMonth: number;
  freeDays: number;
  priceNaira: number;
  includesDelivery: boolean;
}

export interface Meal {
  id: number;
  name: string;
  description: string;
  priceNaira: number;
  imageUrl: string;
  available: boolean;
  /** @nullable */
  category?: string | null;
}

export interface VendorDetail {
  id: number;
  businessName: string;
  area: string;
  cuisineType: string;
  coverImage: string;
  rating: number;
  subscriberCount: number;
  description: string;
  plans: SubscriptionPlan[];
  meals: Meal[];
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone: string;
  area: string;
}

export interface UserProfileUpdate {
  name?: string;
  phone?: string;
  area?: string;
}

export type UserSubscriptionStatus = typeof UserSubscriptionStatus[keyof typeof UserSubscriptionStatus];


export const UserSubscriptionStatus = {
  active: 'active',
  paused: 'paused',
  cancelled: 'cancelled',
} as const;

export interface UserSubscription {
  id: number;
  vendorId: number;
  vendorName: string;
  /** @nullable */
  vendorCoverImage?: string | null;
  planName: string;
  daysPerMonth: number;
  freeDays: number;
  priceNaira: number;
  startDate: string;
  status: UserSubscriptionStatus;
}

export interface SubscriptionInput {
  vendorId: number;
  planId: number;
}

export interface VendorProfile {
  id: number;
  businessName: string;
  ownerName: string;
  email: string;
  phone: string;
  area: string;
  cuisineType: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImage?: string | null;
}

export interface VendorProfileUpdate {
  businessName?: string;
  ownerName?: string;
  phone?: string;
  area?: string;
  cuisineType?: string;
  description?: string;
}

export interface PlanBreakdown {
  planName: string;
  subscriberCount: number;
  revenueNaira: number;
}

export interface VendorDashboard {
  totalSubscribers: number;
  activeSubscribers: number;
  projectedWeeklyEarning: number;
  projectedMonthlyEarning: number;
  totalMeals: number;
  planBreakdown: PlanBreakdown[];
}

export type VendorCustomerStatus = typeof VendorCustomerStatus[keyof typeof VendorCustomerStatus];


export const VendorCustomerStatus = {
  active: 'active',
  paused: 'paused',
  cancelled: 'cancelled',
} as const;

export interface VendorCustomer {
  id: number;
  name: string;
  phone?: string;
  planName: string;
  startDate: string;
  status: VendorCustomerStatus;
  priceNaira: number;
}

export interface VendorEarnings {
  weekly: number;
  monthly: number;
  weeklyByPlan: PlanBreakdown[];
  monthlyByPlan: PlanBreakdown[];
}

export interface MealInput {
  name: string;
  description: string;
  priceNaira: number;
  imageUrl: string;
  category?: string;
  available?: boolean;
}

export interface MealUpdate {
  name?: string;
  description?: string;
  priceNaira?: number;
  imageUrl?: string;
  category?: string;
  available?: boolean;
}

export type ScheduleDayStatus = typeof ScheduleDayStatus[keyof typeof ScheduleDayStatus];


export const ScheduleDayStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
} as const;

export interface ScheduleDay {
  id: number;
  dayNumber: number;
  scheduledDate: string;
  status: ScheduleDayStatus;
  /** @nullable */
  confirmedAt: string | null;
}

export interface WalletWithdrawal {
  id: number;
  amountNaira: number;
  createdAt: string;
}

export interface VendorWallet {
  /** Total earned to date from confirmed pickups */
  earnedNaira: number;
  /** Amount currently available to withdraw */
  withdrawableNaira: number;
  withdrawnNaira: number;
  withdrawals: WalletWithdrawal[];
}

export interface WithdrawInput {
  /** @minimum 1 */
  amountNaira: number;
}

export interface AdminStats {
  vendorCount: number;
  customerCount: number;
  activeSubscriptions: number;
  totalMonthlyRevenueNaira: number;
}

export interface AdminVendor {
  id: number;
  businessName: string;
  ownerName: string;
  email: string;
  phone: string;
  area: string;
  cuisineType: string;
  subscriberCount: number;
}

export interface AdminVendorInput {
  businessName: string;
  ownerName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  area: string;
  cuisineType: string;
  description?: string;
}

export interface AdminCustomer {
  id: number;
  name: string;
  email: string;
  phone: string;
  area: string;
  activeSubscriptionCount: number;
}

export interface AdminCustomerInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  area: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  author: string;
  publishedAt: string;
  coverImage: string;
  /** @nullable */
  category?: string | null;
}

export type ContactInputType = typeof ContactInputType[keyof typeof ContactInputType];


export const ContactInputType = {
  general: 'general',
  support: 'support',
  partnership: 'partnership',
} as const;

export interface ContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
  type?: ContactInputType;
}

export type ListVendorsParams = {
/**
 * Lagos area/neighborhood filter (e.g. Lekki, VI, Ikeja)
 */
area?: string;
/**
 * Search by vendor name or cuisine
 */
search?: string;
};