# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chop Plan API - prepaid lunch subscription service for Lagos restaurants
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user/signup:
    post:
      operationId: userSignup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/user/login:
    post:
      operationId: userLogin
      tags: [auth]
      summary: Login as a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/vendor/signup:
    post:
      operationId: vendorSignup
      tags: [auth]
      summary: Register a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorSignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/vendor/login:
    post:
      operationId: vendorLogin
      tags: [auth]
      summary: Login as a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Login as an admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user or vendor
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── VENDORS (public browsing) ─────────────────────────────────────────────
...
  /admin/vendors:
    get:
      operationId: listAdminVendors
      tags: [admin]
      summary: List all vendors
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminVendor"
    post:
      operationId: createAdminVendor
      tags: [admin]
      summary: Add a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVendor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/vendors/{vendorId}:
    delete:
      operationId: deleteAdminVendor
      tags: [admin]
      summary: Delete a vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/customers:
    get:
      operationId: listAdminCustomers
      tags: [admin]
      summary: List all customers
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCustomer"
    post:
      operationId: createAdminCustomer
      tags: [admin]
      summary: Add a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/customers/{userId}:
    delete:
      operationId: deleteAdminCustomer
      tags: [admin]
      summary: Delete a customer
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── BLOG ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ──────────────────────────────────────────────────────────────
    UserSignupInput:
      type: object
      required: [name, email, password, phone, area]
      properties:
        name:
          type: string
        email:
          type: string

        password:
          type: string
          minLength: 6
        phone:
          type: string
        area:
          type: string
          description: Lagos area/neighborhood
    VendorSignupInput:
      type: object
      required: [businessName, ownerName, email, password, phone, area, cuisineType]
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        email:
          type: string

        password:
          type: string
          minLength: 6
        phone:
          type: string
        area:
          type: string
        cuisineType:
          type: string
        description:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    AuthResponse:
      type: object
      required: [token, role, id, name, email]
      properties:
        token:
          type: string
        role:
          type: string
          enum: [user, vendor, admin]
        id:
          type: integer
        name:
          type: string
        email:
          type: string
    MeResponse:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, vendor, admin]
        area:
          type: ["string", "null"]

    # ── Vendors (public) ──────────────────────────────────────────────────
...
    AdminVendor:
      type: object
      required: [id, businessName, ownerName, email, phone, area, cuisineType, subscriberCount]
      properties:
        id:
          type: integer
        businessName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        area:
          type: string
        cuisineType:
          type: string
        subscriberCount:
          type: integer
    AdminVendorInput:
      type: object
      required: [businessName, ownerName, email, password, phone, area, cuisineType]
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 6
        phone:
          type: string
        area:
          type: string
        cuisineType:
          type: string
        description:
          type: string
    AdminCustomer:
      type: object
      required: [id, name, email, phone, area, activeSubscriptionCount]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        area:
          type: string
        activeSubscriptionCount:
          type: integer
    AdminCustomerInput:
      type: object
      required: [name, email, password, phone, area]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 6
        phone:
          type: string
        area:
          type: string

    # ── Blog ─────────────────────────────────────────────────────────────
...