# lib/api-zod/src/generated/types/adminCustomerInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chop Plan API - prepaid lunch subscription service for Lagos restaurants
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCustomerInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  area: string;
}