# lib/api-zod/src/generated/types/adminVendor.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chop Plan API - prepaid lunch subscription service for Lagos restaurants
 * OpenAPI spec version: 0.1.0
 */

export interface AdminVendor {
  id: number;
  businessName: string;
  ownerName: string;
  email: string;
  phone: string;
  area: string;
  cuisineType: string;
  subscriberCount: number;
}