# lib/api-zod/src/generated/types/adminVendorInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chop Plan API - prepaid lunch subscription service for Lagos restaurants
 * OpenAPI spec version: 0.1.0
 */

export interface AdminVendorInput {
  businessName: string;
  ownerName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  area: string;
  cuisineType: string;
  description?: string;
}